Cleaning 1 and fix bug

+++ Assets/Script/PowerUPPanelle.cs
@@ -12,7 +12,7 @@ public class PowerUPPanelle : MonoBehaviour
     {
         try
         {
-            Text.text = ClientManager.client.allPlayer[PhotonNetwork.LocalPlayer.ActorNumber].var.powerUps[PowerUP].ToString();
+            Text.text = ((int)ClientManager.client.allPlayer[PhotonNetwork.LocalPlayer.ActorNumber].var.powerUps[PowerUP]).ToString();
         }
         catch
         {
